(MODULES-2205) Fix DSC Array generation
When generating the types for each DSC Resource we were not explicitly
making any parameter that was an array in the DSC return an array in
Puppet. Puppet sometimes unwraps arrays if they only contain one item,
which does not work when the DSC Resource parameter is expecting a
string.

To fix this we munge the value to always be an array in DSC Resource type
and make sure that when the PowerShell script is compiled it uses proper
PowerShell array syntax.

Additionaly we change the spec tests to expect array notation if the
parameter is an array