(MODULES-2709) Handle dead process
Previously, if the process died, likely due to when the
process had a malformed command, it would endlessly loop
on stdout being readable (even though there was no output).

Instead, check to see if the process is still alive each time
through the loop and raise an error if the process is dead.